% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densidade.R
\name{densidade}
\alias{densidade}
\title{Gráfico de barras}
\usage{
densidade(data, x_var, y_var)
}
\arguments{
\item{data}{data frame}

\item{x_var}{variável do data frame que ficará no eixo x}

\item{y_var}{variável do data frame que ficará no eixo y}
}
\value{
um gráfico de barras
}
\description{
Esta função retorna um gráfico de barras de acordo com o data frame
passado e duas variáveis escolhidas desse mesmo data frame
}
